Player and Enemy Collision setup
Created a new Scene to work on Player and Enemy collisions

Player script now has a Damage() method, which decreases the player health by 1 when called. There is also a Game Over condition for when health is less than or equal to 0.

Created an EnemyPlayerCollision script which on any collision with the enemy checks to see if the Player was collided with. If so, if calls the Damage method on the Player

Created a canvas and some heart image to represent health