Shield Item
Create a Shield Item. When colliding with the Player, calls a function on the  Player to activate the shield (turns a boolean to true)

If the Player is hit while the boolean is true, they will not be damaged, but the boolean will be set back to false.

There are also two Debug calls to represent this, but will have to be changed to sprite change